Update ness_alarm.markdown
Adding a summary of steps that need to be taken when an IP232 module is installed on the alarm panel. While this information is available in other locations (for example the iComms manual) it is disparate and took me a while to put all the pieces together. My hope is this summary points future users in the right directions much quicker.